Add type Millisecs and use it for interpreting timestmaps
We are moving this type here from https://git.sr.ht/~cryptix/go-toolbelt
which is a module called "go.mindeco.de" for easier maintenance. This
removes dependency on that module. At the same time we are fixing a bug
in this type described in the issue below. Next steps will involve using
this type in other ssb libraries that use this module.

See https://github.com/ssbc/go-ssb/issues/324.